refactor: restructure extension with shared utilities
