+++ .wxt/imports.d.ts
@@ -1,0 +1,6 @@
+export const storage: {
+  defineItem<T>(key: string, opts: { fallback: T }): {
+    getValue(): Promise<T>;
+    setValue(value: T): Promise<void>;
+  };
+};
