+++ README.md
@@ -5,4 +5,17 @@ This template should help get you started developing with Svelte in WXT.
 ## Recommended IDE Setup
 
 [VS Code](https://code.visualstudio.com/) + [Svelte](https://marketplace.visualstudio.com/items?itemName=svelte.svelte-vscode).
+
+## Native Host Setup
+
+This project communicates with a Python native messaging host. To install it:
+
+1. Make the script executable:
+   ```bash
+   chmod +x selector_logger.py
+   ```
+2. Copy `native_host/com.pfahlr.selectorlogger.json` to your browser's native-messaging hosts directory and update the `path` if needed.
+
+The manifest references `../selector_logger.py` so the host can be installed alongside this repository in any location.
+
 # passivepilFRrr
