--- original_project/selector-logger/content.ts
@@ -1,2 +1,0 @@
-// Intentionally empty: we inject code dynamically via popup.ts to ensure
-// freshest logic per click. This file only exists so executeScript can load it once.
