--- original_project/selector-logger/native_host/com.pfahlr.selectorlogger.json
@@ -1,8 +1,0 @@
-{
-  "name": "com.pfahlr.selectorlogger",
-  "description": "Native host for Selector Logger",
-  "path": "/home/rick/Development/firefox-plugins/selector-logger/native_host/selector_logger.py",
-  "type": "stdio",
-  "allowed_origins": ["chrome-extension://jppakgbidmimbpnhpmgaakglhddchgmk/"],
-  "allowed_extensions":  ["selectorlogger@pfahlr"]
-}
