+++ package.json
@@ -3,10 +3,13 @@
   "version": "0.1.0",
   "type": "module",
   "scripts": {
+    "dev": "wxt",
     "build": "wxt build",
-    "lint": "eslint .",
+    "zip": "wxt zip",
     "typecheck": "tsc --noEmit",
-    "test": "npm run lint && npm run typecheck"
+    "lint": "eslint .",
+    "format": "eslint . --fix",
+    "test": "pnpm lint && pnpm typecheck"
   },
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "^6.0.0",
