+++ src/content.ts
@@ -1,0 +1,9 @@
+import { runCollectors } from '~/utils/collectors';
+
+chrome.runtime.onMessage.addListener((message, _sender, sendResponse) => {
+  if (message?.type === 'runCollectors') {
+    const rows = message.rows || [];
+    const result = runCollectors(rows, location.href);
+    sendResponse(result);
+  }
+});
